requirements.txt
python-telegram-bot[job-queue,webhooks]==20.7
requests==2.31.0
aiohttp==3.8.4
python-dotenv==1.0.1
python-binance==1.0.29
pytz
dateparser
yfinance
